requirements.txt
Flask==1.1.2
bert-serving-client
bert-serving-server
fire==0.2.1
nltk==3.4.5
python-dotenv==0.13.0
scikit-learn==0.22.2.post1
tensorflow-gpu==1.14
tensorflow-hub==0.7.0
toposort==1.5
